dev-requirements.txt
pytest-ckan
werkzeug==2.1.2
flask==2.1.2